+++ gpxity/backends/__init__.py
@@ -9,3 +9,4 @@ from .server_directory import *
 from .mmt import *
 from .gpsies import *
 from .trackmmt import *
+from .mailer import *